This PR significantly reduces the ubershader size (985kB to 108kB for the SPIR-V ubershader) by doing two things:
1) Deduplicate logic and loads between the various formats and sizes in the TMEM sampling path. A lot of logic for determining the TMEM addresses to load was easily modified to fit all possible format/size combinations.
2) Move the calls to `sampleTextureLevel` and `clampWrapMirrorSample` into for loops to deduplicate their bodies. This is significant because the lowest level texture sampling code was present 24 times in the original shader. This is because there were 4 `clampWrapMirrorSample` calls in `sampleTextureLevel`, three `sampleTextureLevel` calls in `sampleTexture`, and two `sampleTexture` calls in `RasterPS` for a total of 4 * 3 * 2 = 24 copies.

On all drivers (Mesa in particular), this significantly reduces the ubershader PSO building time for RT64. On my laptop with an Intel iGPU with Mesa, RT64 went from never finishing the ubershader PSO (even after giving it 20 minutes to build), to 10 minutes after some of these optimizations, and eventually to effectively instant with these optimizations completed. A side bonus of these optimizations is a reduction in the size of RT64, which resulted in 7MB shaved off a recomp exe.

There is a very minor impact on ubershader performance from this change, but it also results in vastly improved shader link times which makes it more than worth it in my eyes. On my machine in the same scene, I got these frametimes:

D3D12:
* New:
  * Specialized - 3.1ms
  * Ubershaders = 10.0ms
  * HD Textures Specialized - 3.0ms
  * HD Textures Ubershaders - 7.0ms
* Old:
  * Specialized - 3.1ms
  * Ubershaders - 9.2ms
  * HD Textures Specialized - 3.0ms
  * HD Textures Ubershaders - 6.9ms
Vulkan:
* New:
  * Specialized - 3.0ms
  * Ubershaders - 10.7ms
  * HD Textures specialized - 3.1ms
  * HD Textures Ubershaders - 7.9ms
* Old:
  * Specialized - 3.0ms
  * Ubershaders - 10.0ms
  * HD Textures Specialized - 3.1ms
  * HD Textures Ubershaders - 7.8ms

Meanwhile, dxc shader linking time for specialized shaders was reduced from around 300ms to 50ms average per shader.